Stage-1. NullPointerException caused by request.getParameter() also respond 400.

+++ src/main/java/ru/mail/polis/StandaloneServer/StandaloneServer.java
@@ -1,6 +1,10 @@
 package ru.mail.polis.StandaloneServer;
 
 import one.nio.http.*;
+import org.apache.log4j.LogManager;
+import org.apache.log4j.Logger;
+import org.apache.log4j.spi.LoggerFactory;
+import org.jetbrains.annotations.NotNull;
 import ru.mail.polis.KVDao;
 import ru.mail.polis.KVService;
 
@@ -11,9 +15,12 @@ public class StandaloneServer extends HttpServer implements KVService {
     private final String ID_PARAM = "id=";
     private final String ENTITY_PATH = "/v0/entity";
     private final String STATUS_PATH = "/v0/status";
-    private final byte[] EMPTY_BODY = new byte[0];
 
+    @NotNull
     private final KVDao dao;
+    @NotNull
+    private final Logger logger = LogManager.getLogger(StandaloneServer.class);
+
 
     public StandaloneServer(HttpServerConfig config, KVDao dao) throws IOException {
         super(config);
@@ -32,31 +39,30 @@ public class StandaloneServer extends HttpServer implements KVService {
     @Path(ENTITY_PATH)
     public void entity(Request request, HttpSession session) throws IOException {
         try {
-            if (!request.getParameter(ID_PARAM).isEmpty()) {
-                switch (request.getMethod()) {
-                    case Request.METHOD_PUT: {
-                        dao.upsert(request.getParameter(ID_PARAM).getBytes(), request.getBody());
-                        session.sendResponse(new Response(Response.CREATED, EMPTY_BODY));
-                        System.out.println(Response.CREATED);
-                        return;
+            if (request.getParameter(ID_PARAM) != null) {
+                if (!request.getParameter(ID_PARAM).isEmpty()) {
+                    switch (request.getMethod()) {
+                        case Request.METHOD_PUT: {
+                            dao.upsert(request.getParameter(ID_PARAM).getBytes(), request.getBody());
+                            session.sendResponse(new Response(Response.CREATED, Response.EMPTY));
+                            System.out.println(Response.CREATED);
+                            return;
+                        }
+                        case Request.METHOD_GET: {
+                            session.sendResponse(new Response(Response.OK, dao.get(request.getParameter(ID_PARAM).getBytes())));
+                            return;
+                        }
+                        case Request.METHOD_DELETE: {
+                            dao.remove(request.getParameter(ID_PARAM).getBytes());
+                            session.sendResponse(new Response(Response.ACCEPTED, Response.EMPTY));
+                            return;
+                        }
+                        default: {
+                            session.sendError(Response.METHOD_NOT_ALLOWED, null);
+                        }
                     }
-                    case Request.METHOD_GET: {
-                        session.sendResponse(new Response(Response.OK, dao.get(request.getParameter(ID_PARAM).getBytes())));
-                        return;
-                    }
-                    case Request.METHOD_DELETE: {
-                        dao.remove(request.getParameter(ID_PARAM).getBytes());
-                        session.sendResponse(new Response(Response.ACCEPTED, EMPTY_BODY));
-                        return;
-                    }
-                    default: {
-                        session.sendError(Response.METHOD_NOT_ALLOWED, null);
-                    }
-
-                }
-            } else {
-                session.sendError(Response.BAD_REQUEST, null);
-            }
+                } else session.sendError(Response.BAD_REQUEST, null);
+            } else session.sendError(Response.BAD_REQUEST, null);
         }catch (NoSuchElementException nSEE) {
             nSEE.printStackTrace();
             session.sendError(Response.NOT_FOUND, null);
